feat(j-s): move police case files to shared component
